set store for new login method
