+++ src/stores/userStore.js
@@ -1,26 +1,21 @@
-import axios from "axios";
 import { create } from "zustand";
-import { persist, createJSONStorage } from "zustand/middleware";
+import { persist } from "zustand/middleware";
 
-const host = "http://localhost:8001"; //เดี๋ยวไปใส่ env
+const userStore = (set) => ({
+  userData: null,
+  UserAccount: async (token, input) => {
+    const rs = await createUpdateAccount(token, input);
+    console.log("User", rs);
+    set({ userData: rs.data.results });
+  },
+  GetMyAccount: async (token) => {
+    const rs = await getMyAccount(token)
+    console.log('response', rs);
+    set({ userData: rs.data.results})
+  }
+});
 
-const useUserStore = create(
-	persist(
-		(set, get) => ({
-			user: null,
-			token: "",
-			login: async (input) => {
-				const res = await axios.post(`${host}/auth/login`, input); //Please check
-				set({
-					user: res.data?.user,
-					token: res.data?.token,
-				});
-				return res.data;
-			},
-			logout: () => set({ user: null, token: "" }),
-		}),
-		{ name: "state", storage: createJSONStorage(() => localStorage) }
-	)
-);
+// keep data on Local Storage with persist
+const useUserStore = create(persist(userStore, { name: "user" }))
 
-export default useUserStore;
+export default useUserStore
